// RVCell/ClientApp/src/app/models/propertyUnitModel.ts
export class propertyUnitModel {

  AdditionalInformation: string;
  AgeRequirement: number;
  BookingAvailabilityMessage: string
  BrandID: number;
  ChildrenAllowed: boolean
  CreateDate: string;
  Description: string;
  Installation: string;
  LastModifiedByAccountID: number;
  LastUpdateDate: string;
  MaxOccupancy: number;
  Name: string;
  NumberOfBathrooms: number;
  NumberOfBedrooms: number;
  PageName: string;
  PetDeposit: string;
  PetSizeMaxWeight: number;
  PetsAllowed: boolean;
  PropertyCode: string
  PropertyID: number;
  PropertyName: string;
  PropertyTitle: string;
  PropertyUnitID: number;
  RollawaysAllowed: boolean;
  SquareFootage: string;
  Timestamp: string

}